refactor(dashboard): separate failed queue responsibilities
