+++ compensation/dashboard/src/features/Failed/FailedView.tsx
@@ -11,96 +11,18 @@
  * limitations under the License.
  */
 
-import { useDebouncedFetcherQuery } from "@ahoo-wang/fetcher-react";
-import type { PagedList, PagedQuery } from "@ahoo-wang/fetcher-wow";
-import {
-  all,
-  and,
-  desc,
-  pagedList,
-  pagedQuery,
-  type Condition,
-} from "@ahoo-wang/fetcher-wow";
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { useSearchParams } from "react-router";
-import {
-  ExecutionFailedAggregatedFields,
-  type ExecutionFailedState,
-} from "../../generated";
+import { useGlobalDrawer } from "@/components/GlobalDrawer";
 import { useMediaQuery } from "@/hooks/useMediaQuery";
-import {
-  ResizableHandle,
-  ResizablePanel,
-  ResizablePanelGroup,
-} from "@/components/ui/resizable";
-import {
-  Sheet,
-  SheetContent,
-  SheetDescription,
-  SheetHeader,
-  SheetTitle,
-} from "@/components/ui/sheet";
-import { FailedSearch } from "./FailedSearch.tsx";
-import { FailedTable } from "./FailedTable.tsx";
-import { FindCategory } from "./FindCategory.ts";
-import { RetryConditions } from "./RetryConditions.ts";
-import { clearExecutionSelection, selectExecution } from "./selection.ts";
 import { FailedDetails } from "./details/FailedDetails.tsx";
 import { FetchingFailedDetails } from "./details/FetchingFailedDetails.tsx";
-import { useGlobalDrawer } from "@/components/GlobalDrawer";
+import { FailedSearch } from "./FailedSearch.tsx";
+import { FailedTable } from "./FailedTable.tsx";
+import { FailedWorkspace } from "./FailedWorkspace.tsx";
+import type { FindCategory } from "./FindCategory.ts";
+import { useFailedQueueController } from "./useFailedQueueController.ts";
 
 interface FailedViewProps {
   category: FindCategory;
-}
-
-const SPLIT_LAYOUT_KEY = "compensation-dashboard:failed-view-layout";
-const DEFAULT_SPLIT_LAYOUT = {
-  "execution-list": 40,
-  "execution-details": 60,
-};
-type QueryTransition = "replace" | "pagination" | "refresh";
-
-const executionFailedSort = () => [
-  desc(ExecutionFailedAggregatedFields.AGGREGATE_ID),
-];
-
-function loadSplitLayout(): Record<string, number> {
-  try {
-    const stored = window.localStorage.getItem(SPLIT_LAYOUT_KEY);
-    if (!stored) {
-      return DEFAULT_SPLIT_LAYOUT;
-    }
-    const layout = JSON.parse(stored) as Record<string, unknown>;
-    const list = layout["execution-list"];
-    const details = layout["execution-details"];
-    if (
-      typeof list !== "number" ||
-      typeof details !== "number" ||
-      list < 30 ||
-      list > 52 ||
-      Math.abs(list + details - 100) > 0.1
-    ) {
-      return DEFAULT_SPLIT_LAYOUT;
-    }
-    return { "execution-list": list, "execution-details": details };
-  } catch {
-    return DEFAULT_SPLIT_LAYOUT;
-  }
-}
-
-function saveSplitLayout(layout: Record<string, number>) {
-  try {
-    window.localStorage.setItem(SPLIT_LAYOUT_KEY, JSON.stringify(layout));
-  } catch {
-    // Persisting a UI preference must never interrupt the workspace.
-  }
-}
-
-function isAbortError(error: Error): boolean {
-  return (
-    error.name === "AbortError" ||
-    error.message.toLowerCase().includes("signal is aborted")
-  );
 }
 
 function EmptyDetails() {
@@ -138,221 +60,11 @@ function LoadingPageDetails() {
 export default function FailedView({ category }: FailedViewProps) {
   const desktop = useMediaQuery("(min-width: 960px)");
   const { isOpen: isDrawerOpen } = useGlobalDrawer();
-  const mobileDetailsFocusRef = useRef<HTMLDivElement>(null);
-  const [searchParams, setSearchParams] = useSearchParams();
-  const selectedId = searchParams.get("id");
-  const [searchCondition, setSearchCondition] = useState<Condition>(() =>
-    all(),
-  );
-  const [hasSearchFilters, setHasSearchFilters] = useState(false);
-  const [searchResetToken, setSearchResetToken] = useState(0);
-  const [splitLayout] = useState(loadSplitLayout);
-  const [queryTransition, setQueryTransition] =
-    useState<QueryTransition>("replace");
-  const [settledPagination, setSettledPagination] = useState({
-    index: 1,
-    size: 10,
-  });
-  const [lastSuccessfulPage, setLastSuccessfulPage] =
-    useState<PagedList<ExecutionFailedState>>();
-  const [query, setCurrentQuery] = useState(() =>
-    pagedQuery({
-      condition: RetryConditions.categoryToCondition(category, Date.now()),
-      sort: executionFailedSort(),
-    }),
-  );
-
-  const {
-    loading,
-    result,
-    error,
-    isPending,
-    setQuery: setRemoteQuery,
-  } = useDebouncedFetcherQuery<PagedQuery, PagedList<ExecutionFailedState>>({
-    url: "/execution_failed/snapshot/paged/state",
-    query,
-    debounce: {
-      delay: 300,
-      leading: true,
-    },
-    autoExecute: true,
-    onSuccess: (nextResult) => {
-      const pagination = query.pagination ?? { index: 1, size: 10 };
-      setLastSuccessfulPage(nextResult);
-      setSettledPagination(pagination);
-      setQueryTransition("replace");
-      const lastPage = Math.max(
-        1,
-        Math.ceil(nextResult.total / pagination.size),
-      );
-      if (pagination.index > lastPage) {
-        updateQuery({
-          ...query,
-          sort: executionFailedSort(),
-          pagination: { ...pagination, index: lastPage },
-        });
-      }
-    },
-  });
-  const transitioning = loading || isPending();
-  const visibleError = error && !isAbortError(error) ? error : undefined;
-  const updateQuery = useCallback(
-    (nextQuery: PagedQuery) => {
-      setRemoteQuery(nextQuery);
-      setCurrentQuery(nextQuery);
-    },
-    [setRemoteQuery],
-  );
-
-  const preservingSettledPage =
-    queryTransition !== "replace" &&
-    lastSuccessfulPage !== undefined &&
-    (transitioning || error !== undefined || result === undefined);
-  const suspendingSelection =
-    transitioning &&
-    queryTransition === "pagination" &&
-    lastSuccessfulPage !== undefined;
-  const page =
-    preservingSettledPage && lastSuccessfulPage
-      ? lastSuccessfulPage
-      : transitioning
-        ? pagedList<ExecutionFailedState>()
-        : (result ?? pagedList<ExecutionFailedState>());
-  const staleError = preservingSettledPage ? visibleError : undefined;
-  const blockingError = staleError ? undefined : visibleError;
-  const mutationsDisabled =
-    queryTransition !== "replace" &&
-    (transitioning || visibleError !== undefined);
-  const selectedState = useMemo(() => {
-    if (suspendingSelection) {
-      return undefined;
-    }
-    if (selectedId) {
-      return page.list.find((state) => state.id === selectedId);
-    }
-    return desktop ? page.list[0] : undefined;
-  }, [desktop, page.list, selectedId, suspendingSelection]);
-  const activeId = suspendingSelection
-    ? undefined
-    : (selectedId ?? selectedState?.id);
-  const pagination = query.pagination ?? { index: 1, size: 10 };
-  const pageIndex = pagination.index;
-  const pageSize = pagination.size;
-  const displayedPageIndex = preservingSettledPage
-    ? settledPagination.index
-    : pageIndex;
-  const displayedPageSize = preservingSettledPage
-    ? settledPagination.size
-    : pageSize;
-
-  useEffect(() => {
-    const firstState = page.list[0];
-    if (!desktop || transitioning || selectedId || !firstState) {
-      return;
-    }
-    setSearchParams(selectExecution(searchParams, firstState.id), {
-      replace: true,
-    });
-  }, [
+  const controller = useFailedQueueController({
+    category,
     desktop,
-    page.list,
-    searchParams,
-    selectedId,
-    setSearchParams,
-    transitioning,
-  ]);
-
-  const clearSelection = useCallback(() => {
-    setSearchParams(clearExecutionSelection(searchParams), { replace: true });
-  }, [searchParams, setSearchParams]);
-
-  const onSearch = useCallback(
-    (searchCondition: Condition, hasFilters: boolean) => {
-      setQueryTransition("replace");
-      setLastSuccessfulPage(undefined);
-      setSearchCondition(searchCondition);
-      setHasSearchFilters(hasFilters);
-      clearSelection();
-      updateQuery(
-        pagedQuery({
-          condition: and(
-            RetryConditions.categoryToCondition(category, Date.now()),
-            searchCondition,
-          ),
-          sort: executionFailedSort(),
-        }),
-      );
-    },
-    [category, clearSelection, updateQuery],
-  );
-
-  const clearFilters = useCallback(() => {
-    setSearchResetToken((current) => current + 1);
-    onSearch(all(), false);
-  }, [onSearch]);
-
-  const onPaginationChange = useCallback(
-    (nextPage: number, nextPageSize: number) => {
-      setQueryTransition("pagination");
-      clearSelection();
-      updateQuery({
-        ...query,
-        condition: and(
-          RetryConditions.categoryToCondition(category, Date.now()),
-          searchCondition,
-        ),
-        sort: executionFailedSort(),
-        pagination: { index: nextPage, size: nextPageSize },
-      });
-    },
-    [category, clearSelection, query, searchCondition, updateQuery],
-  );
-
-  const select = useCallback(
-    (state: ExecutionFailedState) => {
-      setSearchParams(selectExecution(searchParams, state.id));
-    },
-    [searchParams, setSearchParams],
-  );
-
-  const refresh = useCallback(() => {
-    setQueryTransition("refresh");
-    updateQuery({
-      ...query,
-      condition: and(
-        RetryConditions.categoryToCondition(category, Date.now()),
-        searchCondition,
-      ),
-      sort: executionFailedSort(),
-    });
-  }, [category, query, searchCondition, updateQuery]);
-
-  useEffect(() => {
-    const timeSensitive = [
-      FindCategory.ToRetry,
-      FindCategory.Executing,
-      FindCategory.NextRetry,
-    ].includes(category);
-    if (!timeSensitive) {
-      return;
-    }
-
-    const refreshVisibleQueue = () => {
-      if (
-        document.visibilityState === "visible" &&
-        !isDrawerOpen &&
-        !transitioning
-      ) {
-        refresh();
-      }
-    };
-    const timer = window.setInterval(refreshVisibleQueue, 30_000);
-    document.addEventListener("visibilitychange", refreshVisibleQueue);
-    return () => {
-      window.clearInterval(timer);
-      document.removeEventListener("visibilitychange", refreshVisibleQueue);
-    };
-  }, [category, isDrawerOpen, refresh, transitioning]);
+    refreshPaused: isDrawerOpen,
+  });
 
   const master = (
     <section
@@ -360,110 +72,53 @@ export default function FailedView({ category }: FailedViewProps) {
       aria-label="Failed executions"
     >
       <FailedSearch
-        key={searchResetToken}
-        onSearch={onSearch}
-        loading={transitioning}
+        key={controller.searchResetToken}
+        onSearch={controller.onSearch}
+        loading={controller.transitioning}
       />
       <FailedTable
-        error={blockingError}
-        hasActiveFilters={hasSearchFilters}
-        loading={transitioning}
-        pagedList={page}
-        pageIndex={displayedPageIndex}
-        pageSize={displayedPageSize}
-        selectedId={activeId}
-        staleError={staleError}
-        onPaginationChange={onPaginationChange}
-        onClearFilters={clearFilters}
-        onRetry={refresh}
-        onSelect={select}
+        error={controller.blockingError}
+        hasActiveFilters={controller.hasSearchFilters}
+        loading={controller.transitioning}
+        pagedList={controller.page}
+        pageIndex={controller.displayedPageIndex}
+        pageSize={controller.displayedPageSize}
+        selectedId={controller.activeId}
+        staleError={controller.staleError}
+        onPaginationChange={controller.onPaginationChange}
+        onClearFilters={controller.clearFilters}
+        onRetry={controller.refresh}
+        onSelect={controller.select}
       />
     </section>
   );
 
-  const details = suspendingSelection ? (
+  const details = controller.suspendingSelection ? (
     <LoadingPageDetails />
-  ) : selectedState ? (
+  ) : controller.selectedState ? (
     <FailedDetails
-      state={selectedState}
-      mutationsDisabled={mutationsDisabled}
-      onChanged={refresh}
+      state={controller.selectedState}
+      mutationsDisabled={controller.mutationsDisabled}
+      onChanged={controller.refresh}
     />
-  ) : selectedId ? (
+  ) : controller.selectedId ? (
     <FetchingFailedDetails
-      key={selectedId}
-      id={selectedId}
-      mutationsDisabled={mutationsDisabled}
-      onChanged={refresh}
+      key={controller.selectedId}
+      id={controller.selectedId}
+      mutationsDisabled={controller.mutationsDisabled}
+      onChanged={controller.refresh}
     />
   ) : (
     <EmptyDetails />
   );
 
-  if (!desktop) {
-    return (
-      <div className="h-full min-h-0">
-        {master}
-        <Sheet
-          open={Boolean(selectedId)}
-          onOpenChange={(open) => {
-            if (!open) {
-              clearSelection();
-            }
-          }}
-        >
-          <SheetContent
-            className="w-full gap-0 p-0 sm:max-w-none"
-            style={{ width: "100%", maxWidth: "none" }}
-            showCloseButton
-            onOpenAutoFocus={(event) => {
-              event.preventDefault();
-              mobileDetailsFocusRef.current?.focus();
-            }}
-          >
-            <SheetHeader className="sr-only">
-              <SheetTitle>Execution failed details</SheetTitle>
-              <SheetDescription>
-                Inspect context and prepare compensation.
-              </SheetDescription>
-            </SheetHeader>
-            <div
-              ref={mobileDetailsFocusRef}
-              tabIndex={-1}
-              aria-label="Execution details panel"
-              className="min-h-0 flex-1 overflow-hidden outline-none"
-            >
-              {details}
-            </div>
-          </SheetContent>
-        </Sheet>
-      </div>
-    );
-  }
-
   return (
-    <ResizablePanelGroup
-      id="failed-executions-layout"
-      orientation="horizontal"
-      className="h-full min-h-0"
-      defaultLayout={splitLayout}
-      onLayoutChanged={(layout, meta) => {
-        if (meta.isUserInteraction) {
-          saveSplitLayout(layout);
-        }
-      }}
-    >
-      <ResizablePanel id="execution-list" minSize="30" maxSize="52">
-        {master}
-      </ResizablePanel>
-      <ResizableHandle
-        withHandle
-        aria-label="Resize execution list and details"
-        className="z-20 bg-slate-200"
-      />
-      <ResizablePanel id="execution-details" minSize="48">
-        <div className="h-full min-h-0">{details}</div>
-      </ResizablePanel>
-    </ResizablePanelGroup>
+    <FailedWorkspace
+      desktop={desktop}
+      details={details}
+      detailsOpen={Boolean(controller.selectedId)}
+      master={master}
+      onCloseDetails={controller.clearSelection}
+    />
   );
 }
